refactor: actualise old solutions  for copy,delte,list, read, rename
