+++ src/fs/copy.js
@@ -1,4 +1,4 @@
-import fs from 'fs';
+import fs from 'fs/promises';
 import path from 'path';
 import { fileURLToPath } from 'url';
 
@@ -10,28 +10,37 @@ const copy = async () => {
     const destinationPath = path.join(__dirname, 'files_copy');
 
     try {
-        if (!fs.existsSync(sourcePath)) {
-            throw new Error('Source folder does not exist');
+        await fs.access(sourcePath);
+        throw new Error('Source folder does not exist');
+    } catch (error) {
+        if (error.code === 'ENOENT') {
+            throw new Error(`FS access operation failed: ${error.message}`);
         }
+    }
 
-        if (fs.existsSync(destinationPath)) {
-            throw new Error('Destination folder already exists');
+    try {
+        await fs.access(destinationPath);
+        throw new Error('Destination folder already exists');
+    } catch (error) {
+        if (error.code === 'ENOENT') {
+            await fs.mkdir(destinationPath);
+        } else {
+            throw new Error(`FS access operation failed: ${error.message}`);
         }
-
-        fs.mkdirSync(destinationPath);
-        const files = fs.readdirSync(sourcePath);
+    }
+    try {
+        const files = await fs.readdir(sourcePath);
 
         for (const file of files) {
             const sourceFile = path.join(sourcePath, file);
             const destinationFile = path.join(destinationPath, file);
-            fs.copyFileSync(sourceFile, destinationFile);
+            await fs.copyFile(sourceFile, destinationFile);
         }
 
         console.log('Folder copied successfully');
     } catch (error) {
         throw new Error(`FS operation failed: ${error.message}`);
     }
-
 };
 
 await copy();
