+++ src/fs/delete.js
@@ -1,4 +1,4 @@
-import fs from 'fs';
+import fs from 'fs/promises';
 import path from 'path';
 import { fileURLToPath } from 'url';
 
@@ -7,18 +7,22 @@ const __dirname = path.dirname(__filename);
 
 const remove = async () => {
     const fileNameToRemove = 'fileToRemove.txt';
+    const filePathToRemove = path.join(__dirname, 'files', fileNameToRemove);
 
     try {
-        const filePathToRemove = path.join(__dirname, 'files', fileNameToRemove);
-
-        if (!fs.existsSync(filePathToRemove)) {
-            throw new Error('File does not exist');
+        await fs.access(filePathToRemove);
+        throw new Error('File does not exist');
+    } catch (error) {
+        if (error.code === 'ENOENT') {
+            throw new Error(`FS access operation failed: ${error.message}`);
         }
+    }
 
-        fs.unlinkSync(filePathToRemove);
+    try {
+        await fs.unlink(filePathToRemove);
         console.log('File deleted successfully');
     } catch (error) {
-        throw new Error(`FS operation failed: ${error.message}`);
+        throw new Error(`FS delete operation failed: ${error.message}`);
     }
 };
