+++ src/fs/rename.js
@@ -1,5 +1,4 @@
-
-import fs from 'fs';
+import fs from 'fs/promises';
 import path from 'path';
 import { fileURLToPath } from 'url';
 
@@ -9,23 +8,28 @@ const __dirname = path.dirname(__filename);
 const rename = async () => {
     const currentFileName = 'wrongFilename.txt';
     const newFileName = 'properFilename.md';
+    const currentFilePath = path.join(__dirname, 'files', currentFileName);
+    const newFilePath = path.join(__dirname, 'files', newFileName);
 
     try {
-        const currentFilePath = path.join(__dirname, 'files', currentFileName);
-        const newFilePath = path.join(__dirname, 'files', newFileName);
-
-        if (!fs.existsSync(currentFilePath)) {
-            throw new Error('Source file does not exist');
-        }
-
-        if (fs.existsSync(newFilePath)) {
-            throw new Error('Destination file already exists');
-        }
-
-        fs.renameSync(currentFilePath, newFilePath);
-        console.log('File renamed successfully');
+        await fs.access(currentFilePath);
+        throw new Error('Source file does not exist');
     } catch (error) {
-        throw new Error(`FS operation failed: ${error.message}`);
+        if (error.code === 'ENOENT') {
+            throw new Error(`FS access operation failed: ${error.message}`);
+        }
+    }
+
+    try {
+        await fs.access(newFilePath);
+        throw new Error('Destination file already exists');
+    } catch (error) {
+        if (error.code === 'ENOENT') {
+            await fs.rename(currentFilePath, newFilePath);
+            console.log('File renamed successfully');
+        } else {
+            throw new Error(`FS access operation failed: ${error.message}`);
+        }
     }
 };
